Fixes for GCC 4.7
gcc/d/d-lang.cc
   d_initialize_diagnostics: context->show_caret is not supported
   on gcc 4.7

The GCC 4.8 branch assumes GCC, GDC and the D FE are built as c++
code. For GCC 4.7 GCC is built as C code. For now keep
the separation between D FE (C++), GDC (C/C++) and GCC(C).

To make it easier to keep this branch in sync with the 4.8
branch we should build all GDC / D FE code as C++ code in
the future, but we have to make sure to guard all GCC imports
with extern(C).