Fix cross-native compilers
A cross-native compiler is a native compiler (host==target) which was
built on a different architecture (build != (host == target)).

This is not a cross compiler and should be handled just like a normal
native compiler.

More importantly: Paths must be in sync with gcc/d/Make-lang.in
otherwise
the compiler doesn't find the import files.