Change `const __sync_op_and` to `enum __sync_op_and`
See http://forum.dlang.org/post/kolzqasjupljvqzivtjf@forum.dlang.org

> 1. In the data segment there is some source code as ascii text 
from a template in gcc/atomics.d . This is in the actual data 
segment and not in debug info segments and goes into the data 
segment of the executable. I do not see any code using this data. 
Why is this in the executable and is it possible to remove it?