Use new unittest runner
We need to run every module test in a seperate process. Otherwise
certain module tests can mess up other modules and we might produce a
deadlock.

The test runner is a slightly extended version of the upstream test
runner. Upstream simply calls the test runner once per module using
the Makefile. This makes it harder to transfer the test executable
to other systems and execute it there. So we add some addditional
code which iterates all compiled in modules and spawns a new process
for every module test. This does essentially what the upstream
Makefiles do, but this solution produces one self-contained
executable.