Link in libphobos and libdruntime in GDC
We used to ship libgdruntime.a and libgphobos2.a where libgphobos.a
also integrated libgdruntime.a. This way we allow users to link
only agains druntime if they don't need phobos.

If we do the same thing with libphobos users can build libraries
against druntime and phobos or only against druntime. But if
a shared phobos library includes druntime you can't link the
phobos and libdruntime libraries (even indirectly via other
libraries). So instead we ship a phobos only libphobos which
depends on druntime. But then we also have to link applications
against libdruntime, not only phobos.

Example:
             |-libphobos
     |--libb-|-libdruntime
foo -|
     |--liba---libdruntime
     |
     |--libdruntime
     |--libphobos