Bug 242: Do constant propagation of enums when frontend doesn't.
Always generate the constructor for a non-scalar manifest constant on
the off-chance that it is referenced instead of propagated in the AST
provided by the front-end.  Also taking into account that if taking the
address of the enum, as what happens when the manifest is the object of
a method call, that it's DECL_INITIAL (or constructor) is always used
instead.

This is a special case, but the right thing to do anyway.