Improve interaction with the binding_level system.
Helps at least the debug generator to keep track of the correct hierachy and references of all var, function, and type declarations.

Also a step in the direction to improve hints for the backend to remove unreferenced nested symbols entirely.  But for now we add anything that is marked TREE_STATIC.