Also copy TYPE_METHODS in copy_struct function
GCC <= 5 does not clear the TYPE_METHODS list in
build_distinct_type_copy which leads to methods with a wrong CONTEXT.
This causes backend errors in the DWARF output module.

We could just clear the TYPE_METHODS for GCC <= 5 instead but providing
copies of the methods should be better for debug info.