Bug 251: ICE: in dwarf2out_imported_module_or_decl_1, at dwarf2out.c:23802
This occurred in complex cases of selectively importing a class
declaration symbol.  The problem being that the TYPE_STUB_DECL is only
set for the underlying RECORD_TYPE, not the reference type.

In the simple case, this wasn't happening because the ImportVisitor
never built an imported decl for straightforward class imports.  This
has also been fixed, which in turn improves gdb usability.

Source file has been reformatted also, with copyright years updated as
it is a relatively manageable size.